Revert "Add validation for wash photo downloads"
